Refactored recipeDetail functions from App.jsx to a custom hook
