+++ src/App.jsx
@@ -3,24 +3,31 @@ import './App.css';
 
 import axios from 'axios';
 
+import PulseLoader from 'react-spinners/PulseLoader';
 import Header from './components/header/Header';
 import SummaryDetail from './components/summaryDetail/SummaryDetail';
 import BackgroundBlur from './components/backgroundBlur/BackgroundBlur';
 import RecipeContainer from './components/recipeDisplay/RecipeContainer';
+import InstructionMenu from './components/instructionMenu/InstructionMenu';
 import SearchBar from './components/searchBar/SearchBar';
 import Footer from './components/footer/Footer';
 import StatusMessage from './components/statusMessage/StatusMessage';
 import ErrorMessage from './components/statusMessage/ErrorMessage';
-import PulseLoader from 'react-spinners/PulseLoader';
+import InstructionMenuHook from './hooks/InstructionMenuHook';
+import RecipeDetailHook from './hooks/RecipeDetailHook';
 
 function App() {
   const [recipes, setRecipes] = useState({});
-  const [recipeDetail, setRecipeDetail] = useState(null);
-  const [isDetailShown, setIsDetailShown] = useState(false);
   const [isSearched, setIsSearched] = useState(false);
   const [isSearching, setIsSearching] = useState(false);
   const [hasError, setHasError] = useState(false);
-
+  const { isInstructionMenuOpen, toggleInstructionMenu } = InstructionMenuHook();
+  const {
+    recipeDetail,
+    isDetailShown,
+    handleRecipeCardClick,
+    handleSummaryDetailClose,
+  } = RecipeDetailHook();
 
   async function fetchData(ingredientString) {
     setIsSearched(true);
@@ -51,21 +58,18 @@ function App() {
     }
     setIsSearching(false);
   }
-
-  function handleRecipeCardClick(recipe) {
-    setRecipeDetail(recipe);
-    setIsDetailShown(true);
-  }
-
-  function handleSummaryDetailClose() {
-    setIsDetailShown(false);
-  }
-
   return (
     <div className="relative flex flex-col min-h-screen">
       <div className="flex-grow">
         <Header />
-        <SearchBar fetchData={fetchData} />
+        <InstructionMenu
+          toggleInstructionMenu={toggleInstructionMenu}
+          isInstructionMenuOpen={isInstructionMenuOpen}
+        />
+        <SearchBar
+          fetchData={fetchData}
+          isInstructionMenuOpen={isInstructionMenuOpen}
+        />
         {isSearching && (
           <div className="flex items-center justify-center p-10">
             <PulseLoader
